requirements.txt
pillow
pandas
numpy
opencv-python
matplotlib
bert-score
inflect
datasets
transformers
seaborn
pymatting
scikit-image
accelerate>=0.26.0
torch
torchvision
torchaudio
av
open-clip-torch==2.24.0
xformers==0.0.29.post2
bitsandbytes==0.45.1